Move ColorIcon to its own directory
