+++ packages/hourly-chart/components/ColorIcon/index.jsx
